boyer moore strnstr: fix pointer arithmetic unsigned overflow
If there's a match of a one or two character pattern at the beginning of
the string, pos ends up being -1 despite being an unsigned size_t. This
makes the undefined sanitizer in clang 5.0.0 complain:

  % echo 'foo bar' | ./ag fo
  ../src/util.c:187:22: runtime error: addition of unsigned offset to 0x557335584520 overflowed to 0x55733558451f
  foo bar

Fix this by making pos a signed ssize_t since it might become negative.
Cast s_len to an ssize_t when comparing to pos to fix a "comparison of
integers of different signs" warning

+++ src/util.c
@@ -178,9 +178,9 @@ void generate_hash(const char *find, const size_t f_len, uint8_t *h_table, const
 const char *boyer_moore_strnstr(const char *s, const char *find, const size_t s_len, const size_t f_len,
                                 const size_t alpha_skip_lookup[], const size_t *find_skip_lookup, const int case_insensitive) {
     ssize_t i;
-    size_t pos = f_len - 1;
+    ssize_t pos = f_len - 1;
 
-    while (pos < s_len) {
+    while (pos < (ssize_t)s_len) {
         for (i = f_len - 1; i >= 0 && (case_insensitive ? tolower(s[pos]) : s[pos]) == find[i]; pos--, i--) {
         }
         if (i < 0) {
